# docs/api/openapi/auth-service.yaml
openapi: 3.1.0
info:
  title: Auth Service API
  version: 1.0.0
  description: Manages user registration, authentication, and JWT lifecycle.
servers:
  - url: https://api.ecommerce.com/api/v1/auth
tags:
  - name: Authentication
    description: User login and registration
paths:
  /register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - firstName
                - lastName
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                firstName:
                  type: string
                  example: John
                lastName:
                  type: string
                  example: Doe
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Validation Error
        '409':
          description: Email already in use
  /login:
    post:
      tags:
        - Authentication
      summary: Authenticate a user and receive JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Successfully authenticated
          headers:
            Set-Cookie:
              description: Refresh token cookie (HttpOnly)
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
        '401':
          description: Invalid credentials
  /refresh:
    post:
      tags:
        - Authentication
      summary: Refresh the access token using HttpOnly cookie
      responses:
        '200':
          description: Successfully refreshed token
        '401':
          description: Invalid or expired refresh token